% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.mcmc.bghs}
\alias{as.mcmc.bghs}
\title{As.mcmc Method for BGHS Objects}
\usage{
as.mcmc.bghs(x, ...)
}
\arguments{
\item{x}{Object of class \code{bghs}.}

\item{...}{Additional arguments.}
}
\description{
As.mcmc Method for BGHS Objects
}
