% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.mcmc.gigg}
\alias{as.mcmc.gigg}
\title{As.mcmc Method for GIGG Objects}
\usage{
as.mcmc.gigg(x, ...)
}
\arguments{
\item{x}{Object of class \code{gigg}.}

\item{...}{Additional arguments.}
}
\description{
As.mcmc Method for GIGG Objects
}
