% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.mcmc.grasp}
\alias{as.mcmc.grasp}
\title{As.mcmc Method for GRASP Objects}
\usage{
as.mcmc.grasp(x, ...)
}
\arguments{
\item{x}{Object of class \code{grasp}.}

\item{...}{Additional arguments.}
}
\description{
As.mcmc Method for GRASP Objects
}
