% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.bghs}
\alias{coef.bghs}
\title{Coef Method for BGHS Objects}
\usage{
\method{coef}{bghs}(object, ...)
}
\arguments{
\item{object}{Object of class \code{bghs}.}

\item{...}{Additional arguments.}
}
\description{
Coef Method for BGHS Objects
}
