% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.gigg}
\alias{coef.gigg}
\title{Coef Method for GIGG Objects}
\usage{
\method{coef}{gigg}(object, ...)
}
\arguments{
\item{object}{Object of class \code{gigg}.}

\item{...}{Additional arguments.}
}
\description{
Coef Method for GIGG Objects
}
