% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.grasp}
\alias{coef.grasp}
\title{Coef Method for GRASP Objects}
\usage{
\method{coef}{grasp}(object, ...)
}
\arguments{
\item{object}{Object of class \code{grasp}.}

\item{...}{Additional arguments.}
}
\description{
Coef Method for GRASP Objects
}
