% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.bghs}
\alias{fitted.bghs}
\title{Fitted Method for BGHS Objects}
\usage{
\method{fitted}{bghs}(object, ...)
}
\arguments{
\item{object}{Object of class \code{bghs}.}

\item{...}{Additional arguments.}
}
\description{
Fitted Method for BGHS Objects
}
