% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.gigg}
\alias{fitted.gigg}
\title{Fitted Method for GIGG Objects}
\usage{
\method{fitted}{gigg}(object, ...)
}
\arguments{
\item{object}{Object of class \code{gigg}.}

\item{...}{Additional arguments.}
}
\description{
Fitted Method for GIGG Objects
}
