% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.grasp}
\alias{fitted.grasp}
\title{Fitted Method for GRASP Objects}
\usage{
\method{fitted}{grasp}(object, ...)
}
\arguments{
\item{object}{Object of class \code{grasp}.}

\item{...}{Additional arguments.}
}
\description{
Fitted Method for GRASP Objects
}
