% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.bghs}
\alias{plot.bghs}
\title{Plot Method for BGHS Objects}
\usage{
\method{plot}{bghs}(x, type = c("interval", "trace", "density"), ...)
}
\arguments{
\item{x}{Object of class \code{bghs}.}

\item{type}{"interval", "trace", or "density".}

\item{...}{Additional arguments.}
}
\description{
Plot Method for BGHS Objects
}
