% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.gigg}
\alias{plot.gigg}
\title{Plot Method for GIGG Objects}
\usage{
\method{plot}{gigg}(x, type = c("interval", "trace", "density"), ...)
}
\arguments{
\item{x}{Object of class \code{gigg}.}

\item{type}{"interval", "trace", or "density".}

\item{...}{Additional arguments.}
}
\description{
Plot Method for GIGG Objects
}
