% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.grasp}
\alias{plot.grasp}
\title{Plot Method for GRASP Objects}
\usage{
\method{plot}{grasp}(x, type = c("interval", "trace", "density"), ...)
}
\arguments{
\item{x}{Object of class \code{grasp}.}

\item{type}{"interval", "trace", or "density".}

\item{...}{Additional arguments.}
}
\description{
Plot Method for GRASP Objects
}
