% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.bghs}
\alias{predict.bghs}
\title{Predict Method for BGHS Objects}
\usage{
\method{predict}{bghs}(object, newdata, type = c("link", "response"), ...)
}
\arguments{
\item{object}{Object of class \code{bghs}.}

\item{newdata}{Data frame for predictions.}

\item{type}{"link" for log-odds or "response" for probabilities.}

\item{...}{Additional arguments.}
}
\description{
Predict Method for BGHS Objects
}
