% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.gigg}
\alias{predict.gigg}
\title{Predict Method for GIGG Objects}
\usage{
\method{predict}{gigg}(object, newdata, type = c("link", "response"), ...)
}
\arguments{
\item{object}{Object of class \code{gigg}.}

\item{newdata}{Data frame for predictions.}

\item{type}{"link" for log-odds or "response" for probabilities.}

\item{...}{Additional arguments.}
}
\description{
Predict Method for GIGG Objects
}
