% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.grasp}
\alias{predict.grasp}
\title{Predict Method for GRASP Objects}
\usage{
\method{predict}{grasp}(object, newdata, type = c("link", "response"), ...)
}
\arguments{
\item{object}{Object of class \code{grasp}.}

\item{newdata}{Data frame for predictions.}

\item{type}{"link" for log-odds or "response" for probabilities.}

\item{...}{Additional arguments.}
}
\description{
Predict Method for GRASP Objects
}
