% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.bghs}
\alias{print.bghs}
\title{Print Method for BGHS Objects}
\usage{
\method{print}{bghs}(x, ...)
}
\arguments{
\item{x}{Object of class \code{bghs}.}

\item{...}{Additional arguments.}
}
\description{
Print Method for BGHS Objects
}
