% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.gigg}
\alias{print.gigg}
\title{Print Method for GIGG Objects}
\usage{
\method{print}{gigg}(x, ...)
}
\arguments{
\item{x}{Object of class \code{gigg}.}

\item{...}{Additional arguments.}
}
\description{
Print Method for GIGG Objects
}
