% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.bghs}
\alias{summary.bghs}
\title{Summary Method for BGHS Objects}
\usage{
\method{summary}{bghs}(object, ...)
}
\arguments{
\item{object}{Object of class \code{bghs}.}

\item{...}{Additional arguments.}
}
\description{
Summary Method for BGHS Objects
}
