% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.gigg}
\alias{summary.gigg}
\title{Summary Method for GIGG Objects}
\usage{
\method{summary}{gigg}(object, ...)
}
\arguments{
\item{object}{Object of class \code{gigg}.}

\item{...}{Additional arguments.}
}
\description{
Summary Method for GIGG Objects
}
