% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.grasp}
\alias{summary.grasp}
\title{Summary Method for GRASP Objects}
\usage{
\method{summary}{grasp}(object, ...)
}
\arguments{
\item{object}{Object of class \code{grasp}.}

\item{...}{Additional arguments.}
}
\description{
Summary Method for GRASP Objects
}
